Initial commit
Here we illustrate some old-school code that even seems well done. It follows good practices and patterns, code is well structured through method codeunits, and it all seems okay. However, the code is tightly coupled and this makes it practically untestable.

This is proven with the test which takes a long time to run. Tests codeunit is absolutely not a best-practice one, it violates some common-sense rules (like not storing authorization keys in code) but all of this is to show how pointless it really is to test any of this business logic, and how impossible it is to test anything in isolation.