Dependency injection introduction: Setup injection
Here we show the simplest of dependency injection scenarios: setup table. Most od BC processes depend on some kind of setup, and that setup is directly retrieved from the database at latest possible moment. However, it makes the code tightly coupled to setup in such a way that the only way to test that code is to first write to setup table (to make sure it contains the necessary setup), and then to run the code that will eventually read from setup table. Finally, when tests are done, the service tier will roll back the changes made to the setup table.

Database access is frequently extremely tightly coupled to business logic in almost all AL code out there. This is so deep in the DNA of AL developers that nobody even considers this.

This demo shows how instead of reading setup at the end of the process, we read it at the beginning and then pass it as a dependency downstream. This allows better testing, because we no longer need to write to the database to test the code: we simply inject the setup record as a dependency. We can still run all our tests, but we don't need to write to the database.