Testing fakes: really important
When providing a fake implementation of an external dependency, it is very important to actually write test code to validate that the fake is properly implemented.

This may sound absurd: writing test code to test other test code. But it is actually what we do here. Bear with us for a moment.

In the previous step we demonstrated that substituting one dependency for another dependency, when done properly, has no effect on code that uses that dependency. We could substitute Azure Key Vault instead of isolated storage, and our code in setup table would still be valid and would still work.

However, to test that, we substituted isolated storage with a fake dependency. The only way to know that the code in setup table still truly works is by making sure that our fake dependency implements its interface correctly. Isolated storage persists data and allows operations on that persisted data. Our fake dependency must also persist data, and we must make sure that it correctly persists that data.

In this example we provide a series of targeted tests where we test all aspects of our fake implementation of `IConfigurationProvider` interface. Those tests prove to us that our fake dependency works exactly as expected, which in turn proves that tests that validate code in setup table after substituting isolated storage with fake, are trustworthy.