Decoupling `send` from `imagine` and `result`
In this commit we decouple even more dependencies. `MidjourneyImagine` and `MidjourneyResult` were tightly coupled to `MidjourneySend` and we could not test them in isolation.

We have now introduced `IMidjourneySend` interface and with it allowed our test code to provide necessary doubles to test `imagine` and `result` endpoint invocations in isolation from actually sending any HTTP requests. The tests focus on properly parsing the responses, which is what these two codeunits really do.

Since we have now injected `IMidjourneySend` as a dependency in `ImagineWithMidjourney` we had to change the signatures of relevant methods in `IMidjourneyImagine` and `IMidjourneyResult` interfaces to accept `IMidjourneySend` dependency. This made us refactor all existing unit tests for `ImagineWithMidjourney` and add pass a `DummySend` as `IMidjourneySend` dependency. This is a dummy double, it does nothing, merely satisfy the requirement for presence of a dependency.