Abstract factory: static
Once the number of dependencies grows, method signatures grow as we are injecting an increasing number of dependencies into methods. Typically, it shouldn't be a problem, but there are also alternatives.

This commit shows an example of an implementation of an abstract factory pattern through a single-instance codeunit. A factory is an object for creating other objects (dependencies). Whenever a piece of code (consumer) requires a specific kind of dependency, that code will ask the factory to provide an instance of that dependency.

Our solution at this point has a number of dependencies, and we can use a factory to instantiate those dependencies when needed.

Test code can preconfigure the factory to return a specific types of dependencies, and this allows us to alter the behavior of code that's being executed by substituting real dependencies with doubles. We achieve the same goal as we did with dependency injection through parameters, except we don't need to pass dependencies into the functions as parameters - instead, we make it functions' responsibility to retrieve their own dependencies from the factory.

This still doesn't violate the DIP (dependency inversion principle) because all consumer code isn't coupled to specific implementations. All consumer code accesses dependencies through their interface, and delegates the responsibility for choosing the right type of the dependency to a factory.

The upside of this approach is that method signatures are simpler, and thus seemingly more readable. We can keep method signatures as simple as they were when everything was tightly coupled, while having the code inside the methods actually decoupled from dependencies.

The downside is that dependency chains are now obscured. With dependency injection as parameters, we knew exactly what dependencies of each process are. Simply by looking at a method's signature, we could get an insight into how the method will perform it's work. With a static factory, this is no longer clear: simply looking at a method's signature we don't know which dependencies it operates upon.

While compiler helped with dependency injection, it no longer helps with static factories: the compiler always knows when you omit a dependency that's expected as a parameter. However, the compiler cannot know if you configured your factory properly before calling a method that will access the factory.