Factory injection
A static factory may be seen as a liability in tests, because a developer could easily forget to properly set it up correctly for a test.

Instead of using a single static factory to replace direct dependency injection through parameters, it's possible to use a combination of dependency injection and factories: factory injection.

This example shows what a factory injection could look like. Instead of depending on a single-instance factory, we abstract the factory itself and allow consumers to pass an explicitly configured instance of a factory to its dependencies.

While this concept still obscures which exact dependencies each component depends on, at least we make it a responsibility of a caller to think of it. By forcing consumers to decide which factory to pass into the methods they call, we make sure that developers never forget that some factory setup may be required.

A further benefit of this approach is that once a factory is configured, it can easily be passed through calls to the entire call stack. When each method that handles dependencies uses a factory to obtain them, and that factory is passed into it as a parameter, then whenever a call is initiated with a specific configuration of a factory, all calls in that call stack will use the same instance of the factory.

An additional benefit is that as more dependencies are added, the only object that has to be changed is factory itself (and its interface, if factory itself implements the factory interface as in our example). Adding more interfaces and their implementations only require to extend the factory, and all components are able to access those new dependencies that were injected into them through factory injection wihtout having to add any extra parameters to existing method signatures. In a world where obsoletion is a major concern for all AL developers, this is an important benefit that can reduce the number of interfaces and their implementations that require obsoletion when they are updated.

Even though this particular example doesn't show this, these dynamic factories are really closest that AL as a language can get to the concept originally described for object-oriented languages.