Wrapping HttpClient dependency
Most of developers will say that "You do not test HTTP calls."

This statement is *almost* true. You don't want your test to make HTTP calls, that part is true. However, you still want to test that your code that interacts with HTTP works correctly.

In a tightly-coupled world of old-school AL, all AL code interacts with `HttpClient` directly and this makes it impossible to test anything that involves `HttpClient`. Developers, even those who routinely write tests for their code, simply skip over writing tests for that.

This is wrong. There are plenty of things you want to test that have to do with HTTP:
* What happens when network is down?
* What happens when calls are blocked by environment? (default behavior for every extension by default in Cloud)
* What if URL moves and is no longer accessible?
* What if your authorization key has expired?
* What if you get invalid JSON response? (this can happen if the API was upgraded)
There are plenty of other scenarios in which you want to run tests against your HTTP code, but without actually using `HttpClient`.

In this commit, we show you how to do that. It's a very large commit with a large number of objects, but in a nutshell, we do this:
* We provide an interface wrapper around `HttpClient`
* We provide a number of handler interfaces for various ways HTTP calls could fail (we focus on three scenarios: transport error, blocked by environment, HTTP error)
* We remove dependency on `HttpClient` and replace it with dependency on our new `IHttpClient` interface

This allows us to write test doubles for all of these interfaces and mock the actual HTTP communication. Our code that consumes HTTP communication has no way of knowing that instead of `HttpClient` tere is a mock framework handling those calls, and we can thoroughly test all things HTTP.

The tests we provide here are of integration kind: we test the entire process, from top to bottom, with all actual dependencies starting from `ImagineWithMidjourney`, through `Imagine` and `Result`, all the way to `Send`. The only thing we substitute with doubles is the HTTP layer.

Enjoy reading the code, and thanks for joining our session and checking out our code repository.